feat!: Extensive rewrite to improve the API
This is primarily related to giving a subscope to elispPackages.
You can now call overrideScope' on elispPackages.

Note on the name of inputOverrides, which was formerly called packageOverrides:
It doesn't have the same format as packageOverrides in Nixpkgs, so it would be
better to use a different name. I am not sure if this is the best name.

Several other minor bugfixes were also made.