fix: Stop trying to parse all library headers
There are some corner cases with parsing library headers.
For example, miniedit.el, which is in emacsorphanage, has a header
named "Author's homepage".

While it is legal to add a library header with space in its name
(see
https://www.gnu.org/software/emacs/manual/html_node/elisp/Library-Headers.html),
it is actually difficult all of the possible corner cases with the
limited capability of Nix regular expressions.
Thus I will give up on parsing all of the library header and
look up individual headers instead, like lisp-mnt.el does.