perf!: Replace subdirs.el with a custom one
This is intended to make the startup process faster by omitting the filesystem
scanning.

By using this function and adding the following setting to your configuration,
the startup becomes faster (0.24 sec -> 0.19 sec for a simple config):

  (setq use-package-ensure-function #'ignore)

This setting is required, because each item in load-path doesn't follow the
convention of package.el.