Make header trimming optional to prevent IFDs
There are some libraries that contain non-ASCII characters, which cannot be
handled in Nix. To work around this issue, an intermediate derivation was
created for each main file, but this is an IFD.

These intermediate derivations come with a performance penalty, it would be
better to avoid them. Not all libraries have the encoding issue, so I made this
optional.

For IFDs, refer to https://nixos.wiki/wiki/Import_From_Derivation