fix: Check for *-pkg.el file in the repository root as well
forge fails to build because of this. It specifies lisp/ in its :files spec and
the root directory is not included in the spec, but forge-pkg.el is in its root.

However, it is still necessary to check *-pkg.el from files in the recipe, which
applies to ghub.el which is a dependency of forge. This shouldn't happen, but I
guess it must be historical.