feat!: Make the library header precede over the *-pkg.el file
This is not a precise implementation of what it supposed to be, but
quite a few packages contain outdated *-pkg.el file.

Especially the version number from the package description file is likely to be
outdated, so it would be better to make it have a lower dependency than the
library header.