feat!: Allow *-pkg.el files without dependencies
Respect dependencies from the package description (*-pkg.el) if and only if the
field exists. Allow descriptions without the field, and, in that case, check the
Package-Requires header.

This is because quite a few *-pkg.el do not specify the dependencies. package.el
works anyway because it is the same as an empty list if it is not defined.

This leads to a bit of nuance. Consider the following case:

1. *-pkg.el specifies nil dependencies or does not contain dependencies.
2. The library header specifies some dependencies, which are outdated.
3. The package actually has no dependencies.

In this case, twist.nix will respect the dependencies from the library header
and build the package with the wrong unnecessary dependencies. If any of the
dependencies is non-existent (e.g. archived), twist simply cannot build the
package. The user must manually overrides packageRequires.