ui: use initReader with initParser as default
The goal here is allowing file-specific behavior in the parser. This is
achieved via an extra optional parameter, initReader. By default, it's
initParser composed with readFile (initParser . readFile, if comfortable
with haskell notation). This makes it compatible with the implementation
prior. If the user needs file-specific treatment, they pass initReader
directly and initParser is not used. Otherwise, setting initParser is
sufficient when all initFiles have equal meaning.