Remove fromElisp input
To simplify maintenance, I have decided to remove flake.lock from this
repository. This is because GitHub Actions are sometimes updated and the change
is propagated to every repository in the organization, which was tedious and
meaningless effort.

In most cases, the user should just use the latest revisions of the
organization. If you don't want to update some of the inputs and encounter an
issue, please file a ticket.